Moved test helpers into `tests/` dir
